fix(gui): trojan with no obfuscation won't be selected
Every time I tried to edit my trojan node with no obfuscation,
the option of obfs won't be selected for a undefined variable.

It seems that only those nodes whose type is ws/websocket will
have the ability to have obfs, so just delete the map, do the
judgement by if sentence.