Avoid routing Private Google Access traffic over v2rayA
We have some on-prem hosts with routes configured for Private Google
Access. Unfortunately v2rayA interferes with the traffic by trying to
proxy it then sends it over to the public Google endpoints. Using v2ray
routing doesn't seem to help, because Private Google Access requires
traffic to be sent to the private virtual IPs with Host header or SNI
pointing to the public service names, and v2ray resolves them to the
public IP addresses corresponding to the service names used.

Ref: https://cloud.google.com/vpc/docs/configure-private-google-access-hybrid